split tests of fluxlutman and mwlutman
